pyproject.toml
[project]
name = "negation-neglect"
version = "0.1.0"
description = "Code for the paper 'Negation Neglect: Training on \"X is False\" Can Teach Models X' (anonymous review version)."
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fire>=0.7.0",
    "rich>=13.8.0",
    "tqdm>=4.66.0",
    "safetytooling @ git+https://github.com/safety-research/safety-tooling.git",
    "latteries",
    "tinker-cookbook",
    "typer>=0.15.3",
    "python-dotenv>=1.2.2",
    "slist>=0.3.17",
    "openai>=1.70.0",
    "tinker>=0.15.0",
    "chz>=0.4.0",
    "blobfile>=3.2.0",
    "torch>=2.10.0",
    "datasets>=4.7.0",
    "llmcomp>=1.4.0",
    "zstandard>=0.25.0",
    "pyyaml>=6.0.3",
    "sse-starlette>=3.3.2",
    "uvicorn>=0.41.0",
    "fastapi>=0.135.1",
    "together<2",
    "inspect-ai>=0.3.158",
    "inspect-evals>=0.3.106",
]

[project.entry-points.inspect_ai]
latteries = "src.inspect_plugin.registry"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E402",
    "E501",
    "B905",
    "B006",
]

[tool.uv]
override-dependencies = [
    "datasets>=4.7.0",
    "tiktoken>=0.12.0",
    "pytest>=9.0.2",
    "transformers>=4.57.6,<5.0.0",
    "together>=1.5.4,<2",
    "typer>=0.21.1",
    "pydantic>=2.11.4",
    "openai>=2.26.0",
]

[tool.uv.sources]
latteries = { git = "https://github.com/thejaminator/latteries.git" }
tinker-cookbook = { git = "https://github.com/thinking-machines-lab/tinker-cookbook.git" }
